requirements.txt
google-adk==0.4.0
google-generativeai==0.8.5
python-dotenv==1.1.0
openai==1.77.0
